Improve warning variety
All warnings are displayed in a similar manner even though some warnings
are more severe than others.

To differentiate between the different types of warnings, a severity
value is added to the warnings, and a custom label to differentiate
between the different warnings.

Addition warnings where the commit subject exceeds 50 characters are
also added to introduce more warning variety. Along with it, an
additional toggle is added to allow the user to customize whether they
need the warning or not.